CRUD Student + Connexion User

+++ src/main/java/eu/ensup/partielspringbootweb/repositories/StudentRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import eu.ensup.partielspringbootweb.entities.Student;
 
 @Repository
-public interface StudentRepository extends CrudRepository<Student, Integer>{
+public interface StudentRepository extends CrudRepository<Student, Long>{
 	
 	public List<Student> findAllByFirstNameAndLastName(String firstName, String lastName);
